+++ CHANGELOG.md
@@ -5,6 +5,7 @@ under development
 -----------------
 - Move contextPool out from router.
 - Add middlewareStatePool.
+- Add Context.DefaultQuery.
 
 v1.9.0 April 4, 2020
 --------------------